Task 1416: Fix Codacy Warnings Package name contains upper case characters
