--- jdi-light-bdd-no-po-tests/src/test/java/cucmberTests/steps/User.java
@@ -1,9 +1,0 @@
-package cucmberTests.steps;
-
-
-import com.epam.jdi.tools.DataClass;
-
-public class User extends DataClass<User> {
-    public String name = "Roman";
-    public String password = "Jdi1234";
-}
