+++ jdi-light-bdd-no-po-tests/src/test/java/cucmbertests/steps/User.java
@@ -1,0 +1,9 @@
+package cucmbertests.steps;
+
+
+import com.epam.jdi.tools.DataClass;
+
+public class User extends DataClass<User> {
+    public String name = "Roman";
+    public String password = "Jdi1234";
+}
